Add some typing. Tidy some imports
