+++ examples/opencv_face_detect_2.py
@@ -1,6 +1,8 @@
 #!/usr/bin/python3
 import time
 
+import time
+
 import cv2
 import numpy as np
